refactor: CTAセクションを共通コンポーネントとして統合
- CTAセクションの重複コードを削除
- 共通コンポーネントを shared/ ディレクトリに移動
- 各ページで共通コンポーネントを再利用するように変更
